Fix content-visibility interaction with style outside flat tree
Non-slotted shadow host children are not part of the flat tree, but we
still allow getComputedStyle() to return styles and such styles are
created on-demand in EnsureComputedStyle() and marked as
IsEnsuredOutsideFlatTree().

These styles were cleared using a kClearEnsured StyleRecalcChange from
shadow hosts during the style recalc pass for subtree recalcs because the
non-slotted elements themselves would not necessarily be marked dirty.
But, there was an assumption that during that step, an element needing
style recalc would need to be cleared since we should have recalculated
styles and marked elements in the flat tree clean before this step.

That was not true for content-visibility which may leave locked subtrees
dirty, which meant we ended up incorrectly recalculating styles inside
locked subtrees.

We already have the detection for clearing ensured computed styles in
EnsureComputedStyle if any element in the ancestor chain is dirty, which
means we can instead mark the direct non-slotted children of the shadow
host for style recalc instead and leave it to EnsureComputedStyle to do
the clearing as necessary.

Bug: 1236774
Change-Id: I3db5fb8f906e703f53abfa2ae8ce86d373a59de2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3114131
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: vmpstr <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#914928}
NOKEYCHECK=True
GitOrigin-RevId: 65c684ca2c45e985b2e8b6079cecf6e0fafe306d