[WebCodecs] Implement support for converting AudioData to float32.
This adds an optional format parameter to the AudioDataCopyToOptions
so that developers can easily interoperate with AudioBuffer. Only
f32-planar is supported at present and a NotSupportedError is thrown
for all other formats.

Conversion is done once into a temporary AudioBus that sticks around
for the lifetime of the AudioData object. We can evaluate switching
to a just-in-time conversion later, but to ensure this can make the
M94 branch, this makes the minimal number of changes required.

R=tguilbert

Bug: 1243034
Change-Id: I31f96f6bc56abe640af92788646ae75c534d6b2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3116565
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>
Cr-Commit-Position: refs/heads/main@{#914962}
NOKEYCHECK=True
GitOrigin-RevId: 44bfca188e7eb0ca5059c63bfc3c5a178a93768a