Fix priority inversion with protobufs initialization on Windows
The lock |mu| is a system-wide lock and we should not hold it while
changing thread priority. The previous attempt to fix this issue
kept the lock to ensure the global variable that boost the
thread priority is not used concurrently.

The current CL is proposing a better design where each thread has
their own instance of a variable of type ScopedBoostThreadPriority.
This is removing the need of the lock to protect access to that
variable. Therefore, we can perform the priority changing outside
of the lock.

This CL is a follow up on:
  https://chromium-review.googlesource.com/c/chromium/src/+/3076381
Example here:
  https://bugs.chromium.org/p/chromium/issues/detail?id=1218253#c10

R=avi@chromium.org, olivierli@chromium.org

Bug: 1218253
Change-Id: I7ae0d630543a687ba1bc069970eab80344392270
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3115052
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Olivier Li <olivierli@chromium.org>
Reviewed-by: Leonard Grey <lgrey@chromium.org>
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#915008}
NOKEYCHECK=True
GitOrigin-RevId: 39bd21e4f0476091e2d18773f10a328118abb44c