Add Blink Extensions to Workers/Worklets.
This CL adds Blink Extensions that are in the ExtensionRegistry
to workers/worklets. There is nothing in this registry unless the
experimental blink_extensions_chromeos GN arg is set, which is off by
default for all configs.

This CL doesn't add tests yet as there're no bots with the gn
configuration. Subsequent CLs will add tests when the GN arg is enabled.

Bug: 1223304
Change-Id: Ife0da6517025e6724d3f0fe3edda932c873d289d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3103248
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Commit-Queue: calamity <calamity@chromium.org>
Cr-Commit-Position: refs/heads/main@{#915041}
NOKEYCHECK=True
GitOrigin-RevId: cf439e7ad7bd45f961be7939028ca5412ea3256a