Force render throttling when clearing prior painted output
LocalFrameView::RenderThrottlingStatusChanged() on the local root
frame checks CanThrottleRendering() to determine whether to clear the
prior paint output, which is correct. But when it runs a paint cycle
to do the actual clearing, ShouldThrottleRendering() can still
return false for various reasons -- for example, if
needs_forced_compositing_update_ is true. In that case, the paint
cycle that is intended only to clear the old paint artifact will
instead do an actual paint.

ForceThrottlingScope is a heavy hammer that supercedes all other
throttling-related logic, and strictly forces all frames in the local
frame tree to be throttled while it is in scope.

Bug: 1239971
Change-Id: I9814b531d386e9b72ebda96eb61a5456b76332d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3119658
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Stefan Zager <szager@chromium.org>
Cr-Commit-Position: refs/heads/main@{#915368}
NOKEYCHECK=True
GitOrigin-RevId: acf6f92d22ef3a71f3df2453c31e968046285bc9