[mathml] Refine definition of embellished operators
In [1], basic logic was added to prepare support for embellished
operators [2]. For example in a LaTeX expression like $u +_V v$,
which may be represented in MathML as

    <math>
      <mi>u</mi>
      <msub><mo>+</mo><mi>V</mi></msub>
      <mi>v</mi>
    </math>

the spacing around the operators should be put around the whole
embellished operator "+ subscript V". Similar behavior is expected
for other operator properties.

In [1], only <mo> elements were considered embellished operators.
This CL extends it to the full definition from MathML Core but
considering only <mtext> and <mspace> as space-like for now [3].
Together, these two CLs add special handling of properties lspace,
rspace and movablelimits with embellished operators. Supporting
other operator properties (e.g. stretchy) and refining the
definition of space-like will happen in follow-up CLs.

MathML Core's definition is based on MathML 3 and is still
relatively complex. Determining whether an element is an
embellished operator and retrieving the property of the core
operator has bad worst-case complexity but it should be fine in
practice [4]. An alternative would be to cache these properties
somewhere in our trees for fast access but it does not seem
worth adding that extra complexity for existing use cases.

Finally, now that mrow-like elements can be embellished operators,
this CL refines the row layout algorithm so that spacing is added
around operator children only if the element being laid out is the
<math> root or is not itself an embellished operator [5]. This
basically addresses the use case of expressions like

    <math>
      <mi>u</mi>
      <mrow><mo>+</mo></mrow>
      <mi>v</mi>
    </math>

where the operator spacing should be added during the layout of the
<math> element, not during the one of the <mrow> element.

This CL improves pass rate for existing WPT tests checking embellished
operator support. Testing is essentially based on lspace/rspace spacing
around embellished operators, as calculated by DetermineOperatorSpacing.
mo-movablelimits-and-embellished-operator.html is added to verify the
case of IsUnderOverLaidOutAsSubSup. This covers the combination of the
change here and in [1].

[1] https://chromium-review.googlesource.com/c/chromium/src/+/3059616
[2] https://w3c.github.io/mathml-core/#embellished-operators
[3] https://w3c.github.io/mathml-core/#dfn-space-like
[4] https://github.com/w3c/mathml/issues/115
[5] https://w3c.github.io/mathml-core/#layout-of-mrow

Bug: 6606, 1124298
Change-Id: I30214e6fadfe4f7deb35b1ebffbc40dbe5023e95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3059456
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Frédéric Wang <fwang@igalia.com>
Cr-Commit-Position: refs/heads/main@{#915472}
NOKEYCHECK=True
GitOrigin-RevId: f7df6b5ef650e99c87c55fbd9651c2d501b9b587