[NFC] CSS highlight painting: clean up selection rect logic
SelectionPaintState calculates the selection rect on demand, caching
the result in a field. This makes sense, because it’s a non-trivial
calculation that isn’t always needed, and the conditions that affect
whether or not it’s needed are complicated.

CL:2647046 added a field that saves the original coordinates when we
rotate the selection rect for a vertical writing-mode, but I did it in
a hacky way that wasn’t very explicit about coordinate spaces. This
patch tries to improve that.

ComputeSelectionRect was also a misleading design, because the offset
arg is only used the first time, and strictly speaking we sometimes
passed coordinates in the wrong space [0]. This patch also eliminates
ComputeSelectionRect in favour of a nullopt check inside the other
selection rect methods, computing the rect if needed using the offset
given in the ctor (which is always in physical space).

Performance impact should be minimal. CurrentLocalSelectionRectForText
is called exactly as often as it used to be, there are two new nullopt
checks per selection painted, and ComputeBoxRect (relatively cheap) is
called only slightly more often than before.

[0] The offset should be in physical space. We pass |box_rect.offset|
at every call site, but after writing-mode rotation, |box_rect| is in
writing-mode space. This merely happens to work because the canvas is
rotated around |box_rect.offset|, so that point is invariant, but it’s
still incorrect. Coordinate handling in text paint code needs systemic
rework to better leverage the type system, or problems like these will
continue to cause pain while we impl compliant highlight painting.

Bug: 1147859
Change-Id: I968726cdf3d73ce9faff6bad8191d2778f46332a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3108999
Commit-Queue: Delan Azabani <dazabani@igalia.com>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#915530}
NOKEYCHECK=True
GitOrigin-RevId: b1fcc4f68ac7fdd9a66e6e88b2b4d3dacd3a7045