Remove !important from <select> color rules
These were added in [1] in response to a bug [2] that was essentially
"you can style the checked option so that it is the same color as the
other options". And while that's true, I'm not sure why that's a good
reason to outlaw changes to color/background-color entirely. As
pointed out in comment [3] there, this CSS today would cause the
same issue:

  option {
    background: rgb(206, 206, 206);
  }

because it would match the existing :checked UA style. And because
of the !important rules, there would be no way to change the :checked
style to ensure contrast.

[1] https://src.chromium.org/viewvc/blink?revision=179782&view=revision
[2] https://crbug.com/398417
[3] https://bugs.chromium.org/p/chromium/issues/detail?id=398417#c8

Fixed: 562990
Bug: 398417
Change-Id: Ic2bf2eaebbe2d736c047e404b855acdcf9c264ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3119649
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#915641}
NOKEYCHECK=True
GitOrigin-RevId: 6fdfd916b7f4ae574ea634dbab6536080633d6f4