[ReplacedNG] Fix saturation issue in ComputeReplacedSize with ratios.
Previously we'd compute "inline_ratio" and "block_ratio". When the
constrained-size was small, and the natural-size large. Both these
ratios would end up being 1/64.

This meant when we compared these ratios, we'd lose the information
that one size was more constrained than the other.

This patch removes the ratios, instead either comparing the sizes
directly (when the ratios where compared to "1") or computing the
hypothetical size in the other axis and comparing the constrained
size to that.

Bug: 1243158
Change-Id: Icac1ce719c5e36acad04f31e5f973e4587e27473
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3123655
Reviewed-by: David Grogan <dgrogan@chromium.org>
Commit-Queue: David Grogan <dgrogan@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#915749}
NOKEYCHECK=True
GitOrigin-RevId: 07251c0764919215304075446d11903dd5ea3e21