[Autofill][Blink] Increased renderer IDs to 64 bit.
This CL increases the the size of FormRendererId and FieldRendererId
from 32 bit to 64 bit.

Integer overflows in 32 bit renderer IDs are feasible within hours,
and would violate Autofill's assumption that renderer IDs uniquely
identify forms and fields.

Details on how Autofill could confuse fields in the case of an integer
overflow are in the bug.

Bug: 1226411
Change-Id: Ia05f0eabb3969fef6913e1615504e64499e7d5fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3000922
Commit-Queue: Christoph Schwering <schwering@google.com>
Reviewed-by: Maria Kazinova <kazinova@google.com>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#915780}
NOKEYCHECK=True
GitOrigin-RevId: 1150f377629ce40679a5d3594ce11a384458d93e