Don't leak underlying MainThreadTaskQueue when MainThreadWebSchedulingTaskQueueImpl is destroyed.
MainThreadWebSchedulingTaskQueueImpl does not own its instance of
MainThreadTaskQueue so when it is destroyed, the MainThreadTaskQueue is
leaked. This CL addresses that by having the WebSchedulingTaskQueue
notify the underlying MainThreadTaskQueue when it is destroyed so the
frame task queue controller can clean up references to and destroy the
MainThreadTaskQueue as well.

This will be removed when crbug.com/1143007 is resolved and the
WebSchedulingTaskQueue owns its MainThreadTaskQueue.

Bug: 1229379
Change-Id: If569ddb2a31fb3b4f10031fea6a5541d8a745b74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3090016
Commit-Queue: Katie Dillon <kdillon@chromium.org>
Reviewed-by: Scott Haseley <shaseley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#915838}
NOKEYCHECK=True
GitOrigin-RevId: 56eda0eca68292755bed42a649c8d768ecbb01d3