Refine counter for BlinkMacSystemFont
Currently use counter for BlinkMacSystemFont is triggered in
StyleBuilderConverterBase as long as that family name is seen in the
font family list. In particular, this counts the typical use case
where authors try to keep backward and forward compatibility:

  font-family:
    ...
    system-ui,          /* new versions */
    BlinkMacSystemFont, /* fallback for old versions */
    ...

Ideally, to get precise metrics we should pass a counter down to macOS's
FontCache::CreateFontPlatformData/MatchNSFontFamily where the font
family system-ui is actually used but this is a bit tedious. Instead,
this CL follows a more simple approach that cover the typical use case
above : only count BlinkMacSystemFont if system-ui was not seen before.
Except for the counter, this does change font-family implementation.

Bug: 554590
Change-Id: Ie2bdb9ba0039e0d8a1fcd9336f2407979a5db41a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3123754
Commit-Queue: Frédéric Wang <fwang@igalia.com>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#915861}
NOKEYCHECK=True
GitOrigin-RevId: 183046ef5e43a9e989fd3762e08fc250e3258d37