[v8] Deserialize code caches off-thread
Add a new feature flag, ConsumeCodeCacheOffThread, which enables
off-thread code cache consumption using a new API added to V8 in
https://crrev.com/c/3067322.

The V8 API follows a similar pattern to the script-streaming API, and
correspondingly the Blink use of this API mirrors Blink's script
streaming support.

At a high level, ScriptResource starts an off-thread code cache
consumption by creating a ScriptCacheConsumer at the moment it receives
a code cache (specifically, when the cache handler is updated to
contain a serialized code cache). The ClassicPendingScript for the
script takes ownership of the ScriptCacheConsumer from the
ScriptResource, and enters an new "waiting for consumer" non-ready state
while it is waiting for the cache consumer to complete. Once the
ScriptCacheConsumer's off-thread deserialization completes, it notifies
the ClassicPendingScript and it becomes "ready", continuing as before.

Once the ClassicPendingScript is ready to execute, it passes ownership
of the ScriptCacheConsumer to ScriptSourceCode. This goes down the
existing paths until V8ScriptRunner, where ownership is passed to V8 via
the v8::ScriptCompiler::Source. From here, V8 takes care of finalising
this off-thread deserialization, and using it instead of the given code
cache data.

ScriptCacheConsumer takes shared ownership of the ScriptResource's
CachedMetadata, to avoid it dying, and state machines on ScriptResource,
ClassicPendingScript, and ScriptCacheConsumer, make sure that their
"finish"/"ready" dance doesn't make a misstep.

Bug: chromium:1075999
Change-Id: I1f29520e43cfdfb47752ca7fc4ed08838d996457
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3067343
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Hiroshige Hayashizaki <hiroshige@chromium.org>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#915890}
NOKEYCHECK=True
GitOrigin-RevId: 8ccb6cf0e030ae981bbf0facc3c70e33e3671e1a