Fix use-after-free errors with ~GPUDevice and device lost handling
Depending on the ordering of ExecutionContext ContextDestroyed
and GC of GPUDevice during page finalization, there can be
use-after-free errors where GPUDevice::lost_callback_ is called after
it has been deleted, or where lost_callback_ is freed twice.

These issues are fixed by setting the C callbacks to null on
~GPUDevice to prevent them from being called after finalization,
and by changing the ownership of lost_callback_ to be a repeating
callback. Repeating callbacks do not self-delete after they are called,
so the only owner of this callback is the GPUDevice.

Further, this CL changes creation of these callbacks so that they can
be used strictly as either repeating callbacks or once callbacks,
making it harder to use them incorrectly.

Lastly, the CL adds a gpu_context_lost regression test which is a stress
test to reproduce the problem discovered in the bug.

Bug: 1242269
Change-Id: Idbad5ba638d221bf1a4b86151e827846cabf4e9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3123966
Commit-Queue: Austin Eng <enga@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#916135}
NOKEYCHECK=True
GitOrigin-RevId: 412b407353a9801468abd04ff43be6e3783a9e0d