Adding flushAndSubmit to allow destruction of textures
After a recent Skia change, Skia's WritePixels will not do the Write
instantly, but wait for a FlushAndSubmit.
Then CanvasResourceProvider creates the SkSurface, backed from the
internal resource.

There is a possibility that the CanvasResourceProvider is destroyed
before the WritePixel skia operation has been executed.

This CL adds a FlushAndSubmit in the destructor, if it's accelerated and
it's not using oop rasterization.

Bug: 1235777
Change-Id: Ie50ccae411626b484d6fd3e6b91e70a480c9faa2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3125134
Reviewed-by: Justin Novosad <junov@chromium.org>
Auto-Submit: Juanmi Huertas <juanmihd@chromium.org>
Commit-Queue: Juanmi Huertas <juanmihd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#916138}
NOKEYCHECK=True
GitOrigin-RevId: 0158c03cee036922c1d01f298009d98f640a5a6d