De-dupe children changed events for misspelled text
Every time a character is typed in a contenteditable, or a navigation
key is pressed, all of the spelling markers are recomputed. Even if
they don't change, they are removed and then added again. In a large
contenteditable, this causes many extra children changed events for
each spelling error.

This changes the AXObjectCacheImpl logic that fires children changed,
to only fire them if the presence of markers has actually changed on
the node.

Technically, it may be possible to fool this if 1 type of
marker is removed and a different type is added. In order to mitigate
this, the de-duping logic that checks for the previously known presence
of markers is only applied to spelling and grammar.

Bug: None
Cq-Include-Trybots: luci.chromium.try:linux-blink-web-tests-force-accessibility-rel
Change-Id: I462c26c9b2e030521d3abd295aeefa43203c6182
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3103692
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#916246}
NOKEYCHECK=True
GitOrigin-RevId: 37eb4c8f8276f2349a2fa197e1704cf86cc01da8