[Conditional Focus][#4] Add tests and remove flag gating
Design-doc:
https://docs.google.com/document/d/1LHJRt-ry9hwzFTbPxKrmD0VvtEFEU6lvqsD7k6wwGKM

Background:
Conditional Focus allows a called to getDisplayMedia() to decide
whether the captured display surface (tab/window) should be focused.
There is a window of 1s during which focus can be triggered or
suppressed. If the application does not explicitly decide within
the allotted time, the captured display surface IS focused,
in line with previous behavior.

This CL:
Introduces browser-tests that verify the feature, then remove
the browser-side gating for the feature. (The feature is still
gated in Blink.)

Next CLs, in order:
1. Change the Blink feature-gating to OT-gating.

Bug: 1215480
Change-Id: I8808dcd32df527d3361756b8621494a843e06779
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3114943
Commit-Queue: Elad Alon <eladalon@chromium.org>
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>
Cr-Commit-Position: refs/heads/main@{#916351}
NOKEYCHECK=True
GitOrigin-RevId: 35203b4496b0940a4759671f8ff91c3298db6668