Don't CreatePattern if there is an issue with the image
When the context is lost, and we try to create a pattern, we can have
some issues with an image created by transferring from an Offscreen
Canvas.

This CL changes a DCHECK for an early return if there is an issue with
the image.

Bug: 1242347
Change-Id: I2f1e9fce75d189fb6ec46f64778fbe2f4d80abe1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3114385
Auto-Submit: Juanmi Huertas <juanmihd@chromium.org>
Commit-Queue: Fernando Serboncini <fserb@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#916442}
NOKEYCHECK=True
GitOrigin-RevId: f9e08c423a840962c0f510c04b02fbaab5eb463b