Avoid breaking violations in nested column rows.
Downgrade the break appeal inside a multicol fragment, if there are
breaking violations, so that we may push the inner column row to the
next outer fragmentainer.

Had to update one test, because we suddenly start to honor orphans and
widows in such situations, which apparently doesn't work in the legacy
fragmentation engine.

Bug: 829028
Change-Id: Ia57fe0c3afa5b16057c0ebdb372653684cd489af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3129467
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#916497}
NOKEYCHECK=True
GitOrigin-RevId: 9a5ca313c038946553086d67aed9ec2845beba9d