[layout] Add another UseCounter for scrollable overflow.
Currently hypothesis is that there are a lot of sites adding a
N+1px x 0px element to a scroller - which doesn't count as scrollable
overflow, but does affect the inflow bounds.

If we add this condition - see how many sites would be effected.

Change-Id: Ib36880e4ffd86f8eff32f3d62f463720130f384b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3131660
Reviewed-by: David Grogan <dgrogan@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#916657}
NOKEYCHECK=True
GitOrigin-RevId: b98e5fb26eade380a72d861ce1e54b17810451de