[Documentation] Add more details to csspaint/README.md
This CL adds some more details to the off-main-thread work flow
for the paint worklet. With this CL, we can easily see the work
flow + its associated code. This documentation will be useful
because we have composite background-color animation, and
clip-path animation, which heavily depends on this infra. We
will likely have more extension in the future using this.

Bug: None
Change-Id: I49a7dd1a22f0b11790b88e70e3f1aa21e5cd4a56
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3133520
Commit-Queue: Xida Chen <xidachen@chromium.org>
Reviewed-by: Yi Gu <yigu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#916884}
NOKEYCHECK=True
GitOrigin-RevId: beaf19cfadf0ad251e8d2bd034b127d91f1e1d0c