Pass playback rate through AnimationEffect::UpdateInheritedTime
Currently, InertEffect effectively has a nullopt playback rate, since
it has no owning Animation to fetch the playback rate from. This causes
InertEffect::IsCurrent to return the wrong answer (e.g. it returns
'false' in the before phase, even when the InertEffect represents
a future Animation object with a playback rate greater than zero).

Instead of pulling it from the owning Animation, we can pass it in.

The playback rate does not appear to be nullable per spec, therefore
UpdateInheritedTime accepts a plain double rather than
optional<double>.

This CL is a preparation for moving the update of the animation
flags [1] to before the animation update is applied.

[1] Currently ElementAnimation::UpdateAnimationFlags

Bug: 1180159
Change-Id: I0a37bb6da6f7ec6f256067350c41279448b4e90b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3122653
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#917185}
NOKEYCHECK=True
GitOrigin-RevId: ca5ca9d9fbe4d8cf6b090b0f10c847b1058db005