Add support for relative URLs in Subresource Loading from WebBundles.
When using a 'resources' attribute in the <link> element with WebBundles
we only use absolute URLs to said resources. This change makes it so we
can use relative URLs instead and resolve them with the page's base URL.
<base> element can cause a dynamic change in the base URL, which we
should support. This will be done in a follow-up CL.

Bug: 1213335, 1244483
Change-Id: I6a2169ba697c102208f7488eef2debc8674cba7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3059820
Commit-Queue: Miras Myrzakerey <myrzakereyms@google.com>
Reviewed-by: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: Hayato Ito <hayato@chromium.org>
Cr-Commit-Position: refs/heads/main@{#917201}
NOKEYCHECK=True
GitOrigin-RevId: 069ca98963b9a33f1dce61a17f94407bff999c09