Fix test picture generation and add test.
This CL changes the test picture from a growing box to a small moving
box. It adds a test that the expected bandwidth target is achieved at
the expected resolution.

In a browser, 640x480 is not reached within 5 seconds. This is odd.

Bug: None.
Change-Id: I2d94f1c190c26cec6759ecc0f788feea8e52ed7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3135833
Commit-Queue: Harald Alvestrand <hta@chromium.org>
Reviewed-by: Florent Castelli <orphis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#917211}
NOKEYCHECK=True
GitOrigin-RevId: 84222e05b7a97c034e40d2d118de4aac05a94c50