Replace forced-break-in-nested-columns.html with a correct WPT test.
The old test was assuming a legacy engine implementation limitation,
where we're unable to push the first fragment of a multicol container to
the next outer fragmentainer if there was no room for its contents in
the current outer fragmentainer.

In the test, see the first DIV child of the inner .mc. The legacy engine
would just push the line inside that DIV to the next outer
fragmentainer, even if there's no valid breakpoint there. The correct
behavior is to push the first fragment of the inner multicol container
itself to the next outer fragmentainer (there's a perfect breakpoint
after the first DIV child of the outer .mc. LayoutNG block fragmentation
gets this right.

Bug: 829028
Change-Id: I55950c5aad3418b49dabe12debd43317cf578c96
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3135656
Reviewed-by: Alison Maher <almaher@microsoft.com>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#917263}
NOKEYCHECK=True
GitOrigin-RevId: 2af5a033f2380e7c2f57a76d4a5a9bd739c55972