[WebID] Change Logout to rely on active session site setting
This causes the Logout API to check for the active session website
setting rather than federated identity request setting.

It also causes the active session setting to be set for a Relying
Party/Identity Provider/Account Identifier triple, in order to enable
Logout for that case.

It also modifies the Logout API to take an Account Identifier so that
the setting can be checked.

Bug: 1223570
Change-Id: I6b54f11f84419d6815035b368ae77221ad1ea837
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3057175
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#917322}
NOKEYCHECK=True
GitOrigin-RevId: a6a6b298350d0d2a1bcc0bd8f0d74e391e6475f8