Do not add test expectations with unknown tag
Due to the discrepancy between CONFIGURATION_SPECIFIER_MACROS and tags
listed in header section of TestExpectations, it is possible we create
an invalid line in TestExpectations when resolve conflicts.

This CL is to make sure we only add test expectations with tag listed in
TestExpectations.

Bug: 1210937
Change-Id: Ib7797e9699e08088d786aae04fab19d4b1923ae2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3138174
Reviewed-by: Dirk Pranke <dpranke@google.com>
Commit-Queue: Weizhong Xia <weizhong@google.com>
Cr-Commit-Position: refs/heads/main@{#917366}
NOKEYCHECK=True
GitOrigin-RevId: 0efd9089dc720d49806acc5a5e7e7ea30a7d98e4