Delay low-priority requests according to the network state in tight mode
Currently, the ResourceLoadScheduler only allows a maximum of two
requests in the tight mode when the priority is lower than high , e.g.
if there is already one running high-priority request, and only another
low-priority one can be handled. This kind of behavior is not
reasonable, because the limit should not be a fixed number, and
we must take the current network state into consideration.

This patch introduces three configurable factors combined together
to decide if a low-priority request will be delayed in the tight mode,
which are:

- Maximum number of the running requests in tight in mode.
- The HTTP RTT threshold.
- The cost reduction when the connections are multiplexed.

The feature is behind
kDelayLowPriorityRequestsAccordingToNetworkState flag,
and two unit tests are added to verify the effect of the factors.

Change-Id: I1001415e86a2b586ec5fdbb8efaaad523b3247f5
Bug: 1130574
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3040720
Commit-Queue: Minggang Wang <minggang@chromium.org>
Reviewed-by: Patrick Meenan <pmeenan@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Cr-Commit-Position: refs/heads/main@{#917713}
NOKEYCHECK=True
GitOrigin-RevId: be61cc9b9c6311f7956f83fc766ccf64c3bb5883