FindInPage: Yield from iterating inner matches.
The find-in-page algorithm currently has yielding behavior
to ensure that we don't start processing a new block if we
already used 10ms of time.

This patch adds a yield from the inner loop as well:
if we processed X entries, then check if we're out of 10ms
budget. If so, early out.

This ensures that in cases we have 100,000s of matches in
a single block (ie large text file), we still yield instead
of blocking until all of the searching is done.

R=yosin@chromium.org, pdr@chromium.org, masonf@chromium.org

Bug: 904480
Change-Id: I04ccb7bd958b7a377a470e9316186d90516375bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3122037
Reviewed-by: Mason Freed <masonf@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Commit-Queue: vmpstr <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#917758}
NOKEYCHECK=True
GitOrigin-RevId: bf30bee1ed35709b3e5ca1bb0ecb7251d7724a09