[Extensions] Specify execution world in scripting.executeScript()
By default, calls to scripting.executeScript() execute in the isolated
world for the extension. However, this is always meant to be a one-way
boundary - the isolated world is always allowed to reach into the main
world and can trivially execute script there by appending a <script>
tag to the DOM. There are many use cases for this, where the script is
intended to be accessed by the page (say, polyfillers).

Allow extensions to explicitly specify the execution world for script
injections by introducing a new enum. The default will remain to
inject in the isolated world, but extensions can choose to inject in
the main world if they desire.

Sample usage:

const tab = getTab();
chrome.scripting.executeScript(
    {
      target: {tabId: tab.id},
      files: ['script.js'],
      world: 'MAIN'
    });

Add browser tests to cover the new functionality.

Bug: 1215518
Change-Id: I7fae6353ba8d00a6f8b6ddf15a536d1404997ba5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3116784
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#917836}
NOKEYCHECK=True
GitOrigin-RevId: 8390ebf91e502183ee3f29cd58b7f1ad6ede3ef3