[WebAudio] Use silence detection for "playback" latency profile
As discussed in the design doc [1], the least invasive approach to
support the usage of FakeAudioWorker for a muted AudioContext is
to check if the current latency profile is "playback".

With this CL, the SilentSinkSuspender will activate the associated
FakeAudioWorker after 30 seconds of silence when the latency
profile is "playback".

This change does not affect other latency profile cases such as
"interactive" and "balanced".

We are adding this feature for the A/B testing, but not for Finch
for the time being. Thus, there's no associated the server-side
configuration.

[1]: go/muted-webaudio

Bug: 1243369
Change-Id: I72e5956fe21ae5e3ea78fdaf37f3f89121063fb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3131282
Reviewed-by: Olga Sharonova <olka@chromium.org>
Commit-Queue: Hongchan Choi <hongchan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#918089}
NOKEYCHECK=True
GitOrigin-RevId: 7187c9b0f10370a505ec6ac9d1ca767e171780aa