Implement one-copy-canvas for WebRTC encoding.
This wires up the RenderableGpuMemoryBufferVideoFramePool for use by
the WebRTC readback path. Frames are read back to GpuMemoryBuffer so
long as they are XRGB/XBGR format.

Unlike elsewhere, WebRTC uses the shared compositor thread context and
doesn't have access to a WebGraphicsContext3DProviderWrapper, so it
can't use WebGraphicsContext3DVideoFramePool directly. As such it
implements its own version of RGMBVFP::Context for use with the
RGMBVFP directly.

This is controlled by the "WebRTCGpuMemoryBufferReadback" feature that
is default on in case we run into any issues with this path.

Bug: 1201009
Change-Id: If052dec871fb3e1afbbfd7266baf88aa121d35a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3025855
Reviewed-by: Ilya Nikolaevskiy <ilnik@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#918213}
NOKEYCHECK=True
GitOrigin-RevId: e4ed74078f4f824dcac7901495add44c87ab269c